plugins-installer: remove hardcoded fallback URL for Flash
Previously, if fetching the .lkgv file from the abacadaba S3 bucket failed, we
would fall back to downloading a hardcoded URL and checking it against a
hardcoded SHA256sum.

Technically, {users who can reach S3} ∩ {users who can reach Adobe's server} ⊆
{users who can reach Adobe's server}, but I think the chances of being able to
reach Adobe's server (currently Akamai's CDN) but not S3 are so slim as to not
be worth worrying about, and so there is little point in maintaining this
fallback data. At best, it's not being used at all; at worst, users will
download an out-of-date copy of Flash.

https://phabricator.endlessm.com/T24616